+++ lib/libs/shared/src/types/loyalty-service/iloyalty-program-sale-channel.d.ts
@@ -2,8 +2,8 @@ import { tags } from 'typia';
 import { BaseResponse } from '../common.type';
 export declare namespace ILoyaltyProgramSaleChannel {
     interface ILoyaltyProgramSaleChannelCreate {
-        sale_channel_id: number & tags.Type<'uint32'>;
-        program_id: number & tags.Type<'uint32'>;
+        sale_channel_id: number & tags.Type<'uint64'>;
+        program_id: number & tags.Type<'uint64'>;
     }
     interface ILoyaltyProgramSaleChannelResponse extends BaseResponse {
         sale_channel_id: number | string;
