+++ lib/libs/shared/src/types/loyalty-service/iloyalty-rule-product.d.ts
@@ -2,8 +2,8 @@ import { tags } from 'typia';
 import { BaseResponse } from '../common.type';
 export declare namespace ILoyaltyRuleProduct {
     interface ILoyaltyRuleProductCreate {
-        product_id: number & tags.Type<'uint32'>;
-        variant_id: number & tags.Type<'uint32'>;
+        product_id: number & tags.Type<'uint64'>;
+        variant_id: number & tags.Type<'uint64'>;
     }
     interface ILoyaltyRuleProductResponse extends BaseResponse {
         rule_id: string | number;
