+++ lib/libs/shared/src/types/payment-service/ipayment-provider.type.d.ts
@@ -1,7 +1,7 @@
 import { tags } from 'typia';
 export declare namespace IPaymentProvider {
     interface IVietQRCreate {
-        transactionId: number & tags.Type<'uint32'>;
+        transactionId: number & tags.Type<'uint64'>;
         bankId: string;
         addInfo: string;
         accountName: string;
